# swagger.yaml
openapi: 3.0.0
info:
  title: Fitness API
  description: |
    API RESTful complète pour l'application mobile Fitness.
    Supporte gestion des utilisateurs, courses, séances, pas, poids, amis, messages et badges.
  version: 1.0.0
  contact:
    name: API Support
    email: support@fitnesssapi.com
  license:
    name: MIT
servers:
  - url: https://fitnesssapi.onrender.com
    description: Production Server
  - url: http://localhost:3000
    description: Local Development Server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentification par Bearer Token (JWT)
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 12
          description: Minimum 12 caracteres avec majuscule, minuscule, chiffre et caractere special.
          example: Motdepasse123!
        name:
          type: string
          minLength: 1
          example: John Doe
        age:
          type: integer
          minimum: 13
          maximum: 120
          example: 30
        heightCm:
          type: number
          maximum: 300
          example: 180
        weightKg:
          type: number
          maximum: 500
          example: 75
        gender:
          type: string
          enum: [male, female, other]
          example: male
        dailyStepsGoal:
          type: integer
          minimum: 1000
          maximum: 100000
          example: 10000
        fitnessLevel:
          type: string
          enum: [beginner, intermediate, advanced]
          example: intermediate
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          example: Motdepasse123!
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT Access Token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: Refresh Token pour renouveller l'access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/UserBasic'

    # ===== User Schemas =====
...
    UserBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        fitnessLevel:
          type: string
        profileImageUrl:
          type: string
          nullable: true
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        body:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        senderId:
          type: string
          format: uuid
        sender:
          $ref: '#/components/schemas/UserBasic'
        workoutId:
          type: string
          format: uuid
          nullable: true
        runningId:
          type: string
          format: uuid
          nullable: true
        workout:
          type: object
          nullable: true
          properties:
            id:
              type: string
            name:
              type: string
            durationMinutes:
              type: integer
            caloriesBurned:
              type: number
        running:
          type: object
          nullable: true
          properties:
            id:
              type: string
            distanceKm:
              type: number
            durationSeconds:
              type: integer
            caloriesBurned:
              type: number
    MessageCreate:
      type: object
      properties:
        body:
          type: string
        workoutId:
          type: string
          format: uuid
        runningId:
          type: string
          format: uuid
...
    MessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        nextCursor:
          type: string
          nullable: true

    # ===== Settings Schemas =====
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Code d'erreur
          example: VALIDATION_ERROR
        message:
          type: string
          description: Message d'erreur lisible
          example: Email or password incorrect
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: S'inscrire
      description: Créer un nouveau compte utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Inscription réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email déjà utilisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Se connecter
      description: Authentification par email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Email ou password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Renouveler le token
      description: Obtenir un nouveau access token avec un refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token renouvelé
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '400':
          description: Token manquant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Refresh token invalide ou expiré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Se déconnecter
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Déconnexion réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
  /auth/google/callback:
    get:
      tags:
        - Authentication
      summary: OAuth Google - Callback
      description: Callback de Google OAuth (redirection)
      parameters:
        - name: code
          in: query
          schema:
            type: string
      responses:
        '303':
          description: Redirection vers deeplink avec tokens

  # ===== USERS =====
...
  /conversations/{id}/messages:
    get:
      tags:
        - Conversations
      summary: Lister les messages
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Messages paginés
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'

    post:
      tags:
        - Conversations
      summary: Envoyer un message
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreate'
      responses:
        '201':
          description: Message envoyé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Contenu nécessaire
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...